feat: [FE/MainPage] 로그인, 위치인증시 화면 렌더링 문제 고침

+++ frontend/src/components/member/logout/Logout.js
@@ -2,7 +2,7 @@ import { Button } from "semantic-ui-react"
 import { useHistory } from "react-router"
 const DEFAULT_REST_URL = process.env.REACT_APP_REST_DEFAULT_URL
 
-const Logout = () => {
+const Logout = ({setIsAuthenticated, setIsCafeAuth}) => {
   const history = useHistory()
   const logout = async () => {
     const response = await fetch(`${DEFAULT_REST_URL}/member/logout`,
@@ -17,11 +17,15 @@ const Logout = () => {
     console.log(responseData)
     if (responseData.httpStatus === "BAD_REQUEST" && responseData.data.sign === "JWT"){
       console.log('logout, 1.유효하지 않은 토큰')
-      history.push('/error')
     }
-    
-    sessionStorage.clear()
-    history.push('/login')
+    else if (responseData.httpStatus==="OK"){
+      setIsAuthenticated(false)
+      setIsCafeAuth(false)
+      sessionStorage.clear()
+      history.push('/login')
+    } else{
+      alert("오류가 발생했습니다.")
+    }
   }
   return <Button onClick={logout}>로그아웃</Button>
 }
